Add files via upload
This is the code I generated in my Youtube videos. The videos are a step by step in generating a .wav file, reading it into to a Python interpreter, displaying the time domain signal, then generating the frequency domain representation, implementing a very crude high-pass filter, then returning the signal to the time domain and writing a new .wav file.